to be squashed; rename to daemon/internal/authconfig
